+++ apps/storefront-e2e/src/support/cy-actions/cart.actions.ts
@@ -6,7 +6,7 @@ declare global {
   namespace Cypress {
     interface Chainable {
       goToCart(): Chainable<void>;
-      goToCartAsGuest(): Chainable<void>;
+      goToGuestCart(): Chainable<void>;
     }
   }
 }
@@ -15,7 +15,7 @@ Cypress.Commands.add('goToCart', () => {
   openCartPage('/customers/DE--**/carts?**');
 });
 
-Cypress.Commands.add('goToCartAsGuest', () => {
+Cypress.Commands.add('goToGuestCart', () => {
   openCartPage('/guest-carts?**');
 });