+++ libs/domain/picking/filters/stories/demo.stories.ts
@@ -1,11 +1,22 @@
-import { OverlaysDecorator } from '@/tools/storybook';
+import {
+  OverlaysDecorator,
+  storybookDefaultViewports,
+} from '@/tools/storybook';
 import { Meta, Story } from '@storybook/web-components';
-import { html, TemplateResult } from 'lit';
+import { TemplateResult, html } from 'lit';
 import { storybookPrefix } from '../../.constants';
 
 export default {
   title: `${storybookPrefix}/Filters`,
-  decorators: [OverlaysDecorator(320, 558)],
+  //900px height is set to be sure that all content fit to the chromatic viewport
+  //that has non-configurable 900px height by default
+  decorators: [OverlaysDecorator(320, 900)],
+  parameters: {
+    layout: 'fullscreen',
+    chromatic: {
+      viewports: [storybookDefaultViewports.mobile.min],
+    },
+  },
 } as Meta;
 
 const Template: Story<unknown> = (): TemplateResult => {