+++ libs/domain/picking/header/stories/demo.stories.ts
@@ -1,9 +1,16 @@
+import { storybookDefaultViewports } from '@/tools/storybook';
 import { Story } from '@storybook/web-components';
 import { html, TemplateResult } from 'lit';
 import { storybookPrefix } from '../../.constants';
 
 export default {
   title: `${storybookPrefix}/Header`,
+  parameters: {
+    layout: 'fullscreen',
+    chromatic: {
+      viewports: [storybookDefaultViewports.mobile.min],
+    },
+  },
 };
 
 const Template: Story = (): TemplateResult => {