+++ libs/domain/picking/user-profile/user-profile.component.ts
@@ -132,7 +132,6 @@ export class UserProfileComponent extends I18nMixin(LitElement) {
             composed: true,
           })
         );
-        this.routerService.navigate('/');
       });
   }