+++ libs/platform/core/src/services/token-resolver/default-token-resolver.service.ts
@@ -1,5 +1,5 @@
 import { resolve } from '@spryker-oryx/di';
-import { map, of, startWith } from 'rxjs';
+import { map, of } from 'rxjs';
 import {
   ResolvedToken,
   TokenResolver,
@@ -28,7 +28,6 @@ export class DefaultTokenService implements TokenResolver {
     }
 
     return tokenResolver.resolve(resolver).pipe(
-      startWith(null),
       map(
         //reverse the value of negative token if needed
         (resolvedValue) => (isNegative ? !resolvedValue : resolvedValue)