+++ tools/storybook/utilities/decorators.ts
@@ -1,40 +1,27 @@
-import { html, LitElement, TemplateResult } from 'lit';
-import { customElement } from 'lit/decorators.js';
+import isChromatic from 'chromatic/isChromatic';
+import { html, TemplateResult } from 'lit';
 import MockDate from 'mockdate';
 
 export const OverlaysDecorator =
-  (minWidth = 1024, minHeight = 768) =>
+  (width = 1024, height = 768) =>
   (storyFn: any, context: any): TemplateResult => {
+    if (!isChromatic()) return storyFn(context);
+
     return html`
-      <div style=${`min-width: ${minWidth}px; min-height: ${minHeight}px;`}>
+      <div style=${`width: ${width}px; height: ${height}px;`}>
         ${storyFn(context)}
       </div>
     `;
   };
 
-@customElement('oryx-mock-date-unmounter')
-class MockDateUnmounter extends LitElement {
-  disconnectedCallback(): void {
-    //allow to reset mocked date when decorated story is disconnected
-    MockDate.reset();
-    super.disconnectedCallback();
-  }
-
-  render() {
-    return html``;
-  }
-}
-
 export const MockDateDecorator =
   (date: string | number | Date = new Date('March 20, 2020 20:00:00')) =>
   (storyFn: any, context: any): TemplateResult => {
-    MockDate.reset();
+    //reset of mocked date happens in global decorator for each story
+    //apps/storybook/.storybook/preview.js
     if (date) {
       MockDate.set(date);
     }
 
-    return html`
-      <oryx-mock-date-unmounter></oryx-mock-date-unmounter>
-      ${storyFn(context)}
-    `;
+    return storyFn(context);
   };